Add SUPER basic conversation agent graph
The agents can now conduct and END a conversation in a basic way. Still
considering moving the user input from a tool that the conversationalist
has to a separate node that the supervisor can call, that way when the
user responds with allergies, the supervisor would detect that and pass
it to the interpreter who would find the matching medicines.

ideas/todo:
  - Fool-proof the convo chain
  - Maybe make user input node
  - Make supervisor/conversationalist realise the user gave
    allergies/diseases so now we pass to the interpreter node
  - something else I forgot.