Intermittent commit, add basic USER_NODE node
This node will replace the getUserInput tool preivously given to the
CONVERSATIONALIST previously, the problem there being that the message
from that was being interpreted as a ToolMessage first and a
HumanMessage second. Not sure if this was a problem but it feels like it
might have come back to bite me. Plus, less responsibility for the
CONVERSATIONALIST.