Add diagnostician agent
So there are cases where we need to identify types of diseases, like
Chron's being a type of malabsorbation disease (that is the name in the
dataframe) or cases where certain medications (astma (if medication
contains glucocorticoids)) interfere with the ECPs.

So we introduce a "guard" agent, which before talking to the interpreter
will judge the set of allergies etc and so forth and if necessary change
out things like 'celiaki' -> 'malabsorbation disease' or ask about
medication and act according to what the answer dictates. E.g. 'customer
has astma' -> 'asthma medicine does not contain glucocorticoids' ->
'asthma is not relevant'.